feat: Add new constants and messages for prompt and chat
