+++ src/constants/message.ts
@@ -1,0 +1,4 @@
+export const templateMessage = {
+  emptyPrompt: "プロンプトが空です。入力してください。",
+  generating: "生成中・・・",
+};
